requirements.txt
RPi.GPIO==0.7.1
Adafruit-DHT==1.4.0
adafruit-circuitpython-ads1x15==2.2.21
board==1.0
busio==5.3.4
Flask==2.3.3
Flask-SocketIO==5.3.6
pandas==2.1.1
numpy==1.24.3
openpyxl==3.1.2
python-socketio==5.8.0
eventlet==0.33.3 
opencv-python==4.8.1.78
face-recognition==1.3.0
dlib==19.24.2
cmake==3.27.7
Pillow==10.0.1
click==8.1.7 